+++ web/src/app/admin/connectors/[connector]/pages/gdrive/Credential.tsx
