--- web/src/app/admin/connectors/[connector]/pages/temporary_drive/google-drive/utils.ts
@@ -1,10 +1,0 @@
-import { GoogleDriveConfig } from "@/lib/ccs/connectors";
-
-export const googleDriveConnectorNameBuilder = (values: GoogleDriveConfig) =>
-  `GoogleDriveConnector-${
-    values.folder_paths && values.folder_paths.join("_")
-  }-${values.include_shared ? "shared" : "not-shared"}-${
-    values.only_org_public ? "org-public" : "all"
-  }-${
-    values.follow_shortcuts ? "follow-shortcuts" : "do-not-follow-shortcuts"
-  }`;
