--- web/src/app/admin/connectors/[connector]/types.ts
@@ -1,1 +1,0 @@
-import { Dispatch, SetStateAction } from "react";
